rewrite Type printer to use typedef heavily
this reworks the CBackend engine to prefer typedef over anonymous types

C does not allow anonymous structs to be compatible with other other
anonymous structs with the same layout, making the heavy use of
them (including the array workaround trick) to be problematic in many
use cases.

instead, this rearranges code emission so that type definitions can be
collected while visiting the functions, then printed (at the top of the
file) at the very end.

this has the additional benefit of making many lines of code more
compact & readable

also fixes an off-by-1 error in the insertvalue instruction emission