Handle simple manipulation of pointers to zero-element arrays (fix #128)
So far, we have translated empty types (such as zero-element arrays) to
C's void type. This works fine if a pointer to such a type is just an
abstract handle, but since you can't do pointer arithmetic on a void
pointer, it does not work for LLVM GEPs.

This commit applies a new strategy for pointers to zero-element arrays.
Instead of treating the array as void, we peel off the array and use
its element type as the pointer target. In other words, we treat
`[Ty x 0]*` as if it were `Ty*`. We can then translate simple GEPs on
such arrays to pointer arithmetic.

The new test (test_empty_array_geps.ll) shows the motivating use-case.